Extracted services and controllers into separate files - Added example service and controller (close #60) - Minor bug fixes - Code cleanup
