+++ src/main/kotlin/ch/uzh/ifi/access/service/ExecutionService.kt
@@ -73,21 +73,22 @@ class ExecutionService(
         val folderId = submission.id.toString() ?: java.util.UUID.randomUUID().toString()
 
         // calculate the embedding in parallel with running the code.
-        val embeddingFuture: CompletableFuture<List<Double>?> = if (isExample(task) && (submission.command == Command.GRADE)) {
-            CompletableFuture.supplyAsync {
-                // Assumption: A submission for an example always consists of only one file, which is the student implementation (to remain language-agnostic
+        val embeddingFuture: CompletableFuture<List<Double>?> =
+            if (isExample(task) && (submission.command == Command.GRADE)) {
+                CompletableFuture.supplyAsync {
+                    // Assumption: A submission for an example always consists of only one file, which is the student implementation (to remain language-agnostic
                     // No testing file for examples / no multi-file examples.
-                if (submission.files.size == 1) {
-                    val implementation = submission.files[0].content ?: ""
-                    courseService.getImplementationEmbedding(implementation)
-                } else {
-                    logger.debug { "More than one file found in the task directory of the submission. It is not clear which file contains the student code." }
-                    null
+                    if (submission.files.size == 1) {
+                        val implementation = submission.files[0].content ?: ""
+                        courseService.getImplementationEmbedding(implementation)
+                    } else {
+                        logger.debug { "More than one file found in the task directory of the submission. It is not clear which file contains the student code." }
+                        null
+                    }
                 }
+            } else {
+                CompletableFuture.completedFuture(null)
             }
-        } else {
-            CompletableFuture.completedFuture(null)
-        }
 
         dockerClient.createContainerCmd(image).use { containerCmd ->
             val submissionDir = workingDir.resolve("submissions").resolve(folderId)
@@ -276,6 +277,7 @@ class ExecutionService(
                 if (results.points != null) {
                     // only relevant for GRADE submissions (RUN and TEST are always valid)
                     submission.valid = true
+                    // TODO: Store test results in submission
                     // never go over 100%; the number of points is otherwise up to the test suite to determine correctly
                     submission.points = minOf(results.points!!, submission.maxPoints!!)
                     evaluation.update(submission.points)
